Adding a battery of unit tests for mir, that should hopefully verify that the test file is being correctly. Also splitting off some tests into a separate module called test_mir_parser, and adding a minor fix to data reader.

+++ pyuvdata/uvdata/mir_parser.py
@@ -750,7 +750,7 @@ class MirParser(object):
         """
         # Gather the needed metadata
         inhid_arr = sp_data["inhid"]
-        nch_arr = sp_data["nch"]
+        nch_arr = sp_data["nch"].astype(np.int)
         dataoff_arr = sp_data["dataoff"] // 2
 
         unique_inhid = np.unique(inhid_arr)
